게시물 목록 조회

+++ src/main/java/com/ssproject/bms/board/repository/BoardRepository.java
@@ -3,5 +3,5 @@ package com.ssproject.bms.board.repository;
 import com.ssproject.bms.board.entity.BoardEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BoardRepository extends JpaRepository<BoardEntity, Long> {
+public interface BoardRepository extends JpaRepository<BoardEntity, Integer> {
 }
